[routing-manager] use a `Tasklet` to re-schedule stale timer
This commit updated how `RoutingManager` re-schedules the stale timer.
It adds a `Tasklet` to calculate the next stale timer and schedules
the timer and changes `ResetDiscoveredPrefixStaleTimer()` to post
this task instead of re-scheduling the timer immediately. This model
allows us to call this method anytime there is a change, (e.g.,
a discovered prefix is added, removed, or updated) and avoid
re-scheduling the timer multiple times when there are a bunch of
related changes (e.g., parsing a RA message containing many RIO/PIO
options updating the discovered prefix list).

This commit also add a call `ResetDiscoveredPrefixStaleTimer()` from
`InvalidateDiscoveredPrefixes()` and `EvaluateOnLinkPrefix()`
where we can potentially remove some entries from the discovered prefix
list and thus require a re-schedule of the stale timer.